% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_oecd_cache}
\alias{clear_oecd_cache}
\title{Clear the readoecd local cache}
\usage{
clear_oecd_cache()
}
\value{
Invisibly returns the number of files deleted.
}
\description{
Deletes all files cached by \code{readoecd} on your local machine. After
clearing, the next call to any \verb{get_oecd_*()} function will re-download
from the OECD API.
}
\examples{
\donttest{
clear_oecd_cache()
}

}
\seealso{
Other utilities: 
\code{\link{check_oecd_api}()},
\code{\link{list_oecd_countries}()}
}
\concept{utilities}
